@@ -26,20 +26,20 @@
 		"@inlang/paraglide-js": "1.11.3",
 		"@poppanator/sveltekit-svg": "5.0.0",
 		"@sveltejs/adapter-auto": "3.3.1",
-		"@sveltejs/kit": "2.12.1",
-		"@sveltejs/vite-plugin-svelte": "5.0.2",
+		"@sveltejs/kit": "2.13.0",
+		"@sveltejs/vite-plugin-svelte": "5.0.3",
 		"@tailwindcss/typography": "0.5.15",
 		"@types/bun": "1.1.14",
 		"@vite-pwa/sveltekit": "0.6.6",
-		"daisyui": "^5.0.0-alpha.50",
+		"daisyui": "^5.0.0-alpha.51",
 		"editorjs-inline-spoiler-tool": "1.0.1",
-		"svelte": "5.14.1",
+		"svelte": "5.15.0",
 		"svelte-check": "4.1.1",
 		"prettier": "3.4.2",
 		"prettier-plugin-svelte": "3.3.2",
 		"sveltekit-superforms": "2.22.1",
 		"typescript": "5.7.2",
-		"vite": "6.0.3",
+		"vite": "6.0.4",
 		"zod": "3.24.1"
 	},
 	"type": "module",
@@ -47,16 +47,16 @@
 		"@editorjs/editorjs": "2.30.7",
 		"@editorjs/marker": "1.4.0",
 		"@iconify/svelte": "4.1.0",
+		"@inlang/paraglide-sveltekit": "0.15.0",
 		"@pilcrowjs/db-query": "0.0.2",
 		"@pilcrowjs/object-parser": "0.0.4",
-		"@inlang/paraglide-sveltekit": "0.15.0",
 		"@svitejs/svelte-preprocess-svg": "2.0.0",
 		"@tailwindcss/vite": "^4.0.0-beta.8",
-		"@tiptap/core": "2.10.3",
-		"@tiptap/pm": "2.10.3",
-		"@tiptap/starter-kit": "2.10.3",
+		"@tiptap/core": "2.10.4",
+		"@tiptap/pm": "2.10.4",
+		"@tiptap/starter-kit": "2.10.4",
 		"arctic": "2.3.3",
-		"cypress": "13.16.1",
+		"cypress": "13.17.0",
 		"editorjs-table": "1.4.10",
 		"editorjs-telegram-post": "1.0.6",
 		"editorjs-undo": "2.0.28",
@@ -65,7 +65,7 @@
 		"panzoom": "9.4.3",
 		"simple-image-editorjs": "1.4.0",
 		"surrealdb": "1.1.0",
-		"svelte-editorjs": "0.1.0",
+		"sveditorjs": "^2.0.1",
 		"svelte-tiptap": "2.1.0",
 		"uuid": "11.0.3"
 	}
